Fixes for Topics with No Image

+++ app.py
@@ -1244,26 +1244,30 @@ def admin_page():
                     topicQuery.name = topicName
 
                     if 'photo' in request.files:
-                        oldImage = None
+                        file = request.files['photo']
+                        if file.filename != '':
+                            oldImage = None
 
-                        if topicQuery.iconClass != None:
-                            oldImage = topicQuery.iconClass
+                            if topicQuery.iconClass != None:
+                                oldImage = topicQuery.iconClass
 
-                        filename = photos.save(request.files['photo'], name=str(uuid.uuid4()) + '.')
-                        topicQuery.iconClass = filename
+                            filename = photos.save(request.files['photo'], name=str(uuid.uuid4()) + '.')
+                            topicQuery.iconClass = filename
 
-                        if oldImage != None:
-                            try:
-                                os.remove(oldImage)
-                            except OSError:
-                                pass
+                            if oldImage != None:
+                                try:
+                                    os.remove(oldImage)
+                                except OSError:
+                                    pass
             else:
                 topicName = request.form['name']
 
                 topicImage = None
                 if 'photo' in request.files:
-                    filename = photos.save(request.files['photo'], name=str(uuid.uuid4()) + '.')
-                    topicImage = filename
+                    file = request.files['photo']
+                    if file.filename != '':
+                        filename = photos.save(request.files['photo'], name=str(uuid.uuid4()) + '.')
+                        topicImage = filename
 
                 newTopic = topics.topics(topicName, topicImage)
                 db.session.add(newTopic)
